Lokmanv2/requirements.txt
# CT Breast Cancer Detection - Production Requirements
# Core ML and Data Processing
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Medical Imaging
pydicom>=2.3.0
SimpleITK>=2.1.0
nibabel>=3.2.0

# Deep Learning Extras
timm>=0.6.0
albumentations>=1.2.0

# Web Interface
flask>=2.2.0
werkzeug>=2.2.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.0.0
pathlib>=1.0.0

# Optional: Experiment Tracking (uncomment if needed)
# wandb>=0.13.0
# tensorboard>=2.10.0

# Development (remove for production)
# jupyter>=1.0.0
# ipykernel>=6.0.0
